refactor: split seats management utilities
