+++ src/components/Seats/SeatsManagement.tsx
@@ -28,287 +28,9 @@ import {
   ExternalLink
 } from 'lucide-react';
 import MapZoomControls from './MapZoomControls';
-import html2canvas from 'html2canvas';
-import jsPDF from 'jspdf';
-
-type PdfMode = 'a4' | 'onePage';
-type ColorMode = 'color' | 'bw';
-
-const DPI_FOR_MM = 96;
-const MM_PER_INCH = 25.4;
-
-function pxToMm(px: number, dpi = DPI_FOR_MM) {
-  return (px / dpi) * MM_PER_INCH;
-}
-
-async function renderWrapperToCanvas(wrapperEl: HTMLElement) {
-  const fullWidth = wrapperEl.scrollWidth || wrapperEl.clientWidth;
-  const fullHeight = wrapperEl.scrollHeight || wrapperEl.clientHeight;
-
-  return await html2canvas(wrapperEl, {
-    width: fullWidth,
-    height: fullHeight,
-    windowWidth: fullWidth,
-    windowHeight: fullHeight,
-    scale: 3,
-    useCORS: true,
-    backgroundColor: '#ffffff',
-  });
-}
-
-function toGrayscaleCanvas(src: HTMLCanvasElement, pureBW = false, threshold = 128) {
-  const w = src.width, h = src.height;
-  const dst = document.createElement('canvas');
-  dst.width = w; dst.height = h;
-  const sctx = src.getContext('2d')!;
-  const dctx = dst.getContext('2d')!;
-  const img = sctx.getImageData(0, 0, w, h);
-  const data = img.data;
-  for (let i = 0; i < data.length; i += 4) {
-    const r = data[i], g = data[i + 1], b = data[i + 2];
-    const y = 0.2126 * r + 0.7152 * g + 0.0722 * b;
-    if (pureBW) {
-      const v = y >= threshold ? 255 : 0;
-      data[i] = data[i + 1] = data[i + 2] = v;
-    } else {
-      data[i] = data[i + 1] = data[i + 2] = y;
-    }
-  }
-  dctx.putImageData(img, 0, 0);
-  return dst;
-}
-
-async function exportMapToPDF(opts: {
-  wrapperEl: HTMLElement;
-  mapLayerEl: HTMLElement;
-  mode: PdfMode;
-  colorMode: ColorMode;
-  bwHard?: boolean;
-  bwThreshold?: number;
-  marginsMm?: number;
-  fileName?: string;
-}) {
-  const {
-    wrapperEl,
-    mapLayerEl,
-    mode,
-    colorMode,
-    bwHard = false,
-    bwThreshold = 128,
-    marginsMm = 10,
-    fileName = 'map.pdf',
-  } = opts;
-
-  mapLayerEl.classList.add('pdf-export');
-  let canvas = await renderWrapperToCanvas(wrapperEl);
-  mapLayerEl.classList.remove('pdf-export');
-
-  if (colorMode === 'bw') {
-    canvas = toGrayscaleCanvas(canvas, bwHard, bwThreshold);
-  }
-
-  if (mode === 'onePage') {
-    const pageWmm = pxToMm(canvas.width);
-    const pageHmm = pxToMm(canvas.height);
-    const orientation = pageWmm > pageHmm ? 'landscape' : 'portrait';
-    const pdf = new jsPDF({ orientation, unit: 'mm', format: [pageWmm, pageHmm], compress: false });
-    pdf.addImage(canvas.toDataURL('image/png'), 'PNG', 0, 0, pageWmm, pageHmm);
-    pdf.save(fileName);
-    return;
-  }
-
-  const pdf = new jsPDF({ orientation: 'portrait', unit: 'mm', format: 'a4', compress: false });
-  const pageW = pdf.internal.pageSize.getWidth();
-  const pageH = pdf.internal.pageSize.getHeight();
-  const maxW = pageW - marginsMm * 2;
-  const maxH = pageH - marginsMm * 2;
-
-  const imgWpx = canvas.width;
-  const imgHpx = canvas.height;
-
-  let renderW = maxW;
-  let renderH = (imgHpx * renderW) / imgWpx;
-  if (renderH > maxH) {
-    renderH = maxH;
-    renderW = (imgWpx * renderH) / imgHpx;
-  }
-  const pxPerMm = imgWpx / renderW;
-  const sliceHeightPx = Math.floor(maxH * pxPerMm);
-
-  for (let yPx = 0; yPx < imgHpx; yPx += sliceHeightPx) {
-    if (yPx > 0) pdf.addPage();
-    const sliceCanvas = document.createElement('canvas');
-    sliceCanvas.width = imgWpx;
-    sliceCanvas.height = Math.min(sliceHeightPx, imgHpx - yPx);
-    sliceCanvas.getContext('2d')!.drawImage(
-      canvas,
-      0,
-      yPx,
-      imgWpx,
-      sliceCanvas.height,
-      0,
-      0,
-      imgWpx,
-      sliceCanvas.height
-    );
-    const sliceHmm = sliceCanvas.height / pxPerMm;
-    pdf.addImage(
-      sliceCanvas.toDataURL('image/png'),
-      'PNG',
-      (pageW - renderW) / 2,
-      marginsMm,
-      renderW,
-      Math.min(sliceHmm, maxH)
-    );
-  }
-
-  pdf.save(fileName);
-}
-
-function PdfToolbar({
-  wrapperRef,
-  mapLayerRef,
-}: {
-  wrapperRef: React.RefObject<HTMLDivElement>;
-  mapLayerRef: React.RefObject<HTMLDivElement>;
-}) {
-  const [pdfMode, setPdfMode] = React.useState<PdfMode>('a4');
-  const [colorMode, setColorMode] = React.useState<ColorMode>('color');
-  const [hardBW, setHardBW] = React.useState(false);
-  const [threshold, setThreshold] = React.useState(128);
-
-  const onExport = async () => {
-    if (!wrapperRef.current || !mapLayerRef.current) return;
-    await exportMapToPDF({
-      wrapperEl: wrapperRef.current,
-      mapLayerEl: mapLayerRef.current,
-      mode: pdfMode,
-      colorMode,
-      bwHard: hardBW,
-      bwThreshold: threshold,
-      marginsMm: 10,
-      fileName:
-        (colorMode === 'bw' ? 'map-bw-' : 'map-color-') +
-        (pdfMode === 'a4' ? 'a4.pdf' : 'onepage.pdf'),
-    });
-  };
-
-  return (
-    <div className="flex flex-wrap items-center gap-3">
-      <label className="text-sm">
-        צבע:
-        <select
-          value={colorMode}
-          onChange={e => setColorMode(e.target.value as ColorMode)}
-          className="ml-2 border rounded px-2 py-1"
-        >
-          <option value="color">צבעוני</option>
-          <option value="bw">שחור-לבן</option>
-        </select>
-      </label>
-
-      {colorMode === 'bw' && (
-        <>
-          <label className="text-sm flex items-center gap-1">
-            <input
-              type="checkbox"
-              checked={hardBW}
-              onChange={e => setHardBW(e.target.checked)}
-            />
-            המרה קשיחה (B/W)
-          </label>
-          {hardBW && (
-            <label className="text-sm">
-              סף:
-              <input
-                type="range"
-                min={0}
-                max={255}
-                value={threshold}
-                onChange={e => setThreshold(+e.target.value)}
-                className="ml-2"
-              />
-              <span className="ml-2">{threshold}</span>
-            </label>
-          )}
-        </>
-      )}
-
-      <label className="text-sm">
-        פורמט:
-        <select
-          value={pdfMode}
-          onChange={e => setPdfMode(e.target.value as PdfMode)}
-          className="ml-2 border rounded px-2 py-1"
-        >
-          <option value="a4">A4 (פיצול עמודים אוטומטי)</option>
-          <option value="onePage">דף אחד (כל המפה)</option>
-        </select>
-      </label>
-
-      <button onClick={onExport} className="px-3 py-1 rounded bg-gray-200 hover:bg-gray-300">
-        ייצוא PDF
-      </button>
-    </div>
-  );
-}
-
-const specialElements = [
-  {
-    id: 'aron-kodesh',
-    name: 'ארון הקודש',
-    type: 'special',
-    width: 120,
-    height: 80,
-    color: '#8B4513',
-    icon: '🕍'
-  },
-  {
-    id: 'bimah',
-    name: 'בימה',
-    type: 'special',
-    width: 100,
-    height: 100,
-    color: '#654321',
-    icon: '📖'
-  },
-  {
-    id: 'amud-tfila',
-    name: 'עמוד תפילה',
-    type: 'special',
-    width: 40,
-    height: 60,
-    color: '#2F4F4F',
-    icon: '🕯️'
-  },
-  {
-    id: 'mizrach',
-    name: 'מזרח',
-    type: 'special',
-    width: 80,
-    height: 40,
-    color: '#FFD700',
-    icon: '🧭'
-  },
-  {
-    id: 'ner-tamid',
-    name: 'נר תמיד',
-    type: 'special',
-    width: 30,
-    height: 30,
-    color: '#FF6347',
-    icon: '🕯️'
-  },
-  {
-    id: 'ezrat-nashim',
-    name: 'עזרת נשים',
-    type: 'special',
-    width: 150,
-    height: 100,
-    color: '#DDA0DD',
-    icon: '👥'
-  }
-];
+import PdfToolbar from './PdfToolbar';
+import { exportMapToPDF } from './pdfUtils';
+import { specialElements } from './specialElements';
 
 const MIN_BENCH_SPACING = 20;
 const MAP_EXPANSION_MARGIN = 20;
