+++ src/components/Seats/specialElements.ts
@@ -1,0 +1,56 @@
+export const specialElements = [
+  {
+    id: 'aron-kodesh',
+    name: 'ארון הקודש',
+    type: 'special',
+    width: 120,
+    height: 80,
+    color: '#8B4513',
+    icon: '🕍'
+  },
+  {
+    id: 'bimah',
+    name: 'בימה',
+    type: 'special',
+    width: 100,
+    height: 100,
+    color: '#654321',
+    icon: '📖'
+  },
+  {
+    id: 'amud-tfila',
+    name: 'עמוד תפילה',
+    type: 'special',
+    width: 40,
+    height: 60,
+    color: '#2F4F4F',
+    icon: '🕯️'
+  },
+  {
+    id: 'mizrach',
+    name: 'מזרח',
+    type: 'special',
+    width: 80,
+    height: 40,
+    color: '#FFD700',
+    icon: '🧭'
+  },
+  {
+    id: 'ner-tamid',
+    name: 'נר תמיד',
+    type: 'special',
+    width: 30,
+    height: 30,
+    color: '#FF6347',
+    icon: '🕯️'
+  },
+  {
+    id: 'ezrat-nashim',
+    name: 'עזרת נשים',
+    type: 'special',
+    width: 150,
+    height: 100,
+    color: '#DDA0DD',
+    icon: '👥'
+  }
+];
